+++ package.json
@@ -37,6 +37,7 @@
     "@types/luxon": "^2.0.0",
     "@types/node": "^16.4.10",
     "@types/nodemailer": "^6.4.14",
+    "@types/nunjucks": "^3.2.6",
     "@types/uuid": "^8.3.1",
     "@types/validator": "^13.6.3",
     "@typescript-eslint/eslint-plugin": "^4.29.0",
@@ -76,6 +77,7 @@
     "meilisearch": "^0.30.0",
     "morgan": "^1.10.0",
     "nodemailer": "^6.9.12",
+    "nunjucks": "^3.2.4",
     "query-string": "^7.0.1",
     "sqlite3": "^5.1.2",
     "uuid": "^8.3.2",
