+++ src/services/index.ts
@@ -3,6 +3,7 @@ export * as bannerService from './banner'
 export * as cacheService from './cache'
 export * as categoryService from './categories'
 export * as favoriteService from './favorite'
+export * as programService from './program'
 export * as searchService from './search'
 export * as trendService from './trends'
 export * as userActivityService from './user-activity'
