+++ templates/_breadcrumbs.html
@@ -1,0 +1,10 @@
+<div style="font-size: 0.9rem; color: #999; padding: 20px 0 5px;">
+  {% for crumb in crumbs %}
+    {% set url = crumb[0] %}
+    {% set label = crumb[1] %}
+    <a href="{{ url }}">{{ label }}</a>
+    {% if crumbs.indexOf(crumb) < crumbs.length - 1 %} 
+      <span style="display: inline-block; padding-left: 5px; padding-right: 5px;">/</span>
+    {% endif %}
+  {% endfor %}
+</div>
