+++ templates/_header.html
@@ -1,0 +1,3 @@
+<nav>
+  <a href="/sitemap"><img src="/logo.svg" alt="AUW Logo" style="height: 50px;" /></a>
+</nav>
