+++ templates/programs.html
@@ -1,0 +1,22 @@
+{% set pageTitle = 'All Programs' %}
+{% set pageDescription = 'Trusted source of social service resources in the state of Hawaiʻi' %}
+{% set pageURL = 'https://search.auw211.org' %}
+{% set pageImage = '/logo.svg' %}
+
+{% extends "_base.html" %}
+
+{% block header %}
+{% include "_header.html" %}
+{% endblock %}
+
+{% block main %}
+
+{% set crumbs = [['/sitemap', 'Home']] %}
+{% include "_breadcrumbs.html" %}
+
+<ul>
+  {% for program in programs %}
+  <li><a href="/sitemap/programs/{{ program.id }}">{{ program.service_name }} at {{ program.location_name }}</a></li>
+  {% endfor %}
+</ul>
+{% endblock %}
